Add packaged app MCP entrypoint
